feat: use autoloader.
I want to know vlucas/phpdotenv can autoloader class? I try it but fail
