+++ composer.json
@@ -1,5 +1,10 @@
 {
     "require": {
         "vlucas/phpdotenv": "^5.6"
+    },
+    "autoload": {
+        "psr-4": {
+            "MyNamespace\\": "src/MyNamespace/"
+        }
     }
 }
