+++ createHero.php
@@ -1,5 +1,4 @@
 <?php
-require_once 'hero.php';
 require_once 'mysql.php';
 // createHero();
