+++ monster.php
@@ -1,5 +1,4 @@
 <?php
-require 'character.php';
 $level_1_monster = new Character('Slime', '');
 $level_1_monster->hp = 30;
 $level_1_monster->printAttributes();
