+++ src/character.class.php
