+++ src/mage.class.php
@@ -1,0 +1,6 @@
+<?php
+
+class Mage extends Hero
+{
+    public $magic_attack = 5;
+}
