+++ src/warrior.class.php
@@ -1,0 +1,6 @@
+<?php
+class Warrior extends Hero
+{
+    public $hp = 20;
+    public $physical_attack = 5;
+}
